test: add manifest snapshots
